refactor: rename create-handler to handler and update related imports
